bubblebot-messenger/requirements.txt
aredis
apscheduler
httpx
pyyaml
pycryptodome
websockets
loguru
typing_extensions
